+++ outputs/all/all.go
@@ -1,6 +1,6 @@
 package all
 
 import (
+	_ "github.com/influxdb/telegraf/outputs/datadog"
 	_ "github.com/influxdb/telegraf/outputs/influxdb"
-	_ "github.com/jipperinbham/telegraf/outputs/datadog"
 )